Fix line ending in test case when running on Windows
This test case executes `cat` and uses `PHP_EOL` as line ending.  This
causes the test to fail when executed on Windows connected to a Linux
SSH server.  Since `cat` is usually not available on Windows, running
the test would fail if a Windows SSH server was used.  Thus, we can
easily fix this by always using LF.