Fix #80729: ssh2_shell() streams are not stream_select()able
On Windows, the socket type is not an `int` but rather `SOCKET` which
is actually `unsigned int` on x86 and `unsigned __int64` on x64.  To
cater to this type mismatch, PHP offers the portable `php_socket_t` (at
least as of PHP 5.0.0).  We use it for `php_ssh2_session_data` as well.